chore: move the MomentoRpcMethod enum back to main dir
