+++ momento-sdk/src/intTest/java/momento/sdk/retry/TestRetryMetricsCollectorTest.java
@@ -4,7 +4,7 @@ import static org.junit.jupiter.api.Assertions.*;
 
 import java.util.List;
 import java.util.Map;
-import momento.sdk.MomentoRpcMethod;
+
 import momento.sdk.TestRetryMetricsCollector;
 import org.junit.jupiter.api.Test;
